Refactor navigation and analytics processing logic
Extracted authentication and settings navigation into separate files (AuthNavigation.kt, SettingsNavigation.kt) and updated AppNavigation.kt to use these nested graphs for improved modularity. Refactored AnalyticsViewModel to encapsulate analytics processing in a new ProcessedAnalytics data class, improving code clarity and separation of concerns.
