+++ app/src/main/java/com/zenia/app/ui/navigation/AppNavigation.kt
@@ -20,49 +20,30 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.zenia.app.ui.screens.MainScreen
-import com.zenia.app.ui.screens.account.AccountRoute
 import com.zenia.app.ui.screens.analytics.AnalyticsRoute
 import com.zenia.app.ui.screens.analytics.AnalyticsViewModel
-import com.zenia.app.ui.screens.auth.AuthRoute
-import com.zenia.app.ui.screens.lock.LockRoute
-import com.zenia.app.ui.screens.auth.AuthViewModel
-import com.zenia.app.ui.screens.auth.ForgotPasswordScreen
 import com.zenia.app.ui.screens.diary.DiarioRoute
 import com.zenia.app.ui.screens.diary.DiaryEntryScreen
+import com.zenia.app.ui.screens.lock.LockRoute
 import com.zenia.app.ui.screens.notifications.NotificationsRoute
-import com.zenia.app.ui.screens.premium.PremiumRoute
 import com.zenia.app.ui.screens.recursos.RecursosRoute
 import com.zenia.app.ui.screens.relax.RelaxRoute
-import com.zenia.app.ui.screens.settings.DonationsRoute
-import com.zenia.app.ui.screens.settings.HelpCenterRoute
-import com.zenia.app.ui.screens.settings.MoreSettingsRoute
-import com.zenia.app.ui.screens.settings.PrivacyRoute
-import com.zenia.app.ui.screens.settings.SettingsRoute
 import com.zenia.app.ui.screens.sos.HelplineRoute
 import com.zenia.app.viewmodel.MainViewModel
-import com.zenia.app.viewmodel.SettingsViewModel
 import kotlinx.coroutines.delay
 import java.time.LocalDate
 
 /**
  * Composable principal que gestiona la navegación de toda la aplicación.
  * Utiliza un [NavHost] para definir todas las rutas (pantallas) posibles.
- *
- * Obtiene los ViewModels de autenticación ([AuthViewModel]) y configuración ([SettingsViewModel])
- * para determinar la pantalla de inicio correcta.
  */
 @Composable
 fun AppNavigation(pendingDeepLink: Uri? = null) {
     val navController = rememberNavController()
 
     val mainViewModel: MainViewModel = hiltViewModel()
-    /**
-     * Lógica clave para determinar la pantalla de inicio de la app (startDestination).
-     * 1. Si el usuario está logueado Y tiene biometría activada -> Va a [Destinations.LOCK_ROUTE].
-     * 2. Si está logueado pero SIN biometría -> Va directo a [Destinations.HOME_ROUTE].
-     * 3. Si no está logueado (en cualquier otro caso) -> Va a [Destinations.AUTH_ROUTE].
-     */
     val startDestination by mainViewModel.startDestinationState.collectAsState()
+
     if (startDestination == null) {
         Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
             CircularProgressIndicator()
@@ -90,25 +71,8 @@ fun AppNavigation(pendingDeepLink: Uri? = null) {
         navController = navController,
         startDestination = startDestination!!
     ) {
-        composable(Destinations.AUTH_ROUTE) {
-            val authViewModel: AuthViewModel = hiltViewModel()
-            AuthRoute(
-                authViewModel = authViewModel,
-                onNavigateToForgotPassword = {
-                    navController.navigate(Destinations.FORGOT_PASSWORD_ROUTE)
-                }
-            )
-        }
-
-        composable(Destinations.FORGOT_PASSWORD_ROUTE) {
-            val authViewModel: AuthViewModel = hiltViewModel()
-            ForgotPasswordScreen(
-                viewModel = authViewModel,
-                onNavigateBack = {
-                    navController.popBackStack()
-                }
-            )
-        }
+        authGraph(navController)
+        settingsGraph(navController, mainViewModel)
 
         composable(
             route = Destinations.HOME_ROUTE,
@@ -170,105 +134,6 @@ fun AppNavigation(pendingDeepLink: Uri? = null) {
                 onNavigateBack = { navController.popBackStack() },
                 onNavigateToDestination = { route ->
                     navController.navigate(route)
-                }
-            )
-        }
-
-        composable(
-            route = Destinations.SETTINGS_ROUTE,
-            enterTransition = { slideIn() },
-            exitTransition = { slideOut() },
-            popEnterTransition = { popSlideIn() },
-            popExitTransition = { popSlideOut() }
-        ) {
-            val settingsVM: SettingsViewModel = hiltViewModel()
-            SettingsRoute(
-                settingsViewModel = settingsVM,
-                onNavigateBack = { navController.popBackStack() },
-                onNavigateToProfile = { navController.safeNavigate(Destinations.ACCOUNT_ROUTE) },
-                onNavigateToPremium = { navController.safeNavigate(Destinations.PREMIUM_ROUTE) },
-                onNavigateToMoreSettings = { navController.safeNavigate(Destinations.MORE_SETTINGS_ROUTE) },
-                onNavigateToHelp = { navController.safeNavigate(Destinations.HELP_CENTER_ROUTE) },
-                onNavigateToDonations = { navController.safeNavigate(Destinations.DONATIONS_ROUTE) },
-                onNavigateToPrivacy = { navController.safeNavigate(Destinations.PRIVACY_POLICY_ROUTE) },
-                onSignOut = {
-                    mainViewModel.signOut()
-                    navController.navigate(Destinations.AUTH_ROUTE) {
-                        popUpTo(Destinations.HOME_ROUTE) { inclusive = true }
-                    }
-                },
-            )
-        }
-
-        composable(
-            route = Destinations.MORE_SETTINGS_ROUTE,
-            enterTransition = { slideIn() },
-            exitTransition = { slideOut() },
-            popEnterTransition = { popSlideIn() },
-            popExitTransition = { popSlideOut() }
-        ) {
-            MoreSettingsRoute(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        composable(
-            route = Destinations.PREMIUM_ROUTE,
-            enterTransition = { slideIn() },
-            exitTransition = { slideOut() },
-            popEnterTransition = { popSlideIn() },
-            popExitTransition = { popSlideOut() }
-        ) {
-            PremiumRoute(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        composable(
-            route = Destinations.HELP_CENTER_ROUTE,
-            enterTransition = { slideIn() },
-            exitTransition = { slideOut() },
-            popEnterTransition = { popSlideIn() },
-            popExitTransition = { popSlideOut() }
-            ) {
-            HelpCenterRoute(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        composable(
-            route = Destinations.DONATIONS_ROUTE,
-            enterTransition = { slideIn() },
-            exitTransition = { slideOut() },
-            popEnterTransition = { popSlideIn() },
-            popExitTransition = { popSlideOut() }
-            ) {
-            DonationsRoute(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        composable(
-            route = Destinations.PRIVACY_POLICY_ROUTE,
-            enterTransition = { slideIn() },
-            exitTransition = { slideOut() },
-            popEnterTransition = { popSlideIn() },
-            popExitTransition = { popSlideOut() }
-            ) {
-            PrivacyRoute(
-                onNavigateBack = { navController.popBackStack() }
-            )
-        }
-
-        composable(Destinations.ACCOUNT_ROUTE) {
-            val authViewModel: AuthViewModel = hiltViewModel()
-            AccountRoute(
-                authViewModel = authViewModel,
-                onNavigateBack = { navController.popBackStack() },
-                onNavigateToAuth = {
-                    navController.navigate(Destinations.AUTH_ROUTE) {
-                        popUpTo(Destinations.HOME_ROUTE) { inclusive = true }
-                    }
                 }
             )
         }
